new function for primality
danaj as found in perlmonks provided a new, faaster function to check primality.
in the below bench the referred as dj2 and is faster:

           Rate original      dj2
original 6.36/s       --     -44%
dj2      11.3/s      78%       --

danaj also suggest to use Math::Prime::Util because ntheory checking will be hundreds of times faster on small numbers, and many thousands of times faster for bigger.
Here is not used because speed is not really a concern. I adopted danaj's pure perl solution as 'quote' because danaj was patient with me and interested in this project.

Speed is not a concern here because.. prime numbers can be found only in the column of natural numbers ie the second and last-1 ones.
So the biggest prime number will be less or equal to the max row of the triangle and we dont want a tringle too big anyway.